[fix]: refactor spring security endpoint
